// src/app/models/usuario.model.ts
export class Usuario {
  idUsuario?: number;
  nombre?: string;
  apellido?: string;
  dni?: number;
  telefono?: number;
  usuario?: string;
  contrasena?: string;
  direccion?: string;
  nombreCompleto?: string;
  usuarioSuperior?: number;
  usuarioRegistro?: number;
  usuarioActualiza?: number;
}